Add a section on how to run under a sub-dir
This adds the start of a section on how to run geohealthcheck under a sub-dir. It's taken me a while to figure out how to do this in my setup without changing any core code, using nginx and gunicorn, and the end solution was so simple that I thought it was worth contributing.  Unfortunately I don't know how it would be done using any other webserver! I do wonder whether I should also contribute a section on how to run geohealthcheck using nginx and gunicorn anyway!