Fix auth encoding/decoding
From code review by @justb4:
6 - Auth info Decoding/Encoding Failure with Postgres

Only occurs when using Postgres, not with SQLite. Background: GHC
encodes/encrypts a generic auth dict info structure via JSON string to be stored
as textfield in DB. It decodes/decrypts when reading. This way we can support
multiple auth types with a single auth column in resource table. So this is
another encoding than in 4 for HTTP auth headers, but think similar problem.
We also need to deal with existing PG DBs that have auth columns already present
in resource table.

-  install psycopg2 : pip install psycopg2
-  used existing PG DB, only changed in config_site.py: SQLALCHEMY_DATABASE_URI
   = 'postgresql://name:passw@localhost:5432/ghc', but may create new
-  add URL Resource with Basic Auth (even does not have to have basic auth)
-  in Edit add Basic Auth Username and Password
-  click Save

Fixed by ensuring encode/decode takes `string` as input and gives `string` as
output. These methodes now use static typing. The origin of the problem was the
difference of how SQLite and Postgres store the encoded string (either as text
or in some binary form). This solution makes the encoded authentication a
string object and *not* bytes.