Update http.py
The Catalogue Service (CSW) with HTTP POST Resource URL with body - Probe doesn't send a Content-Type and generated a 415 (Unsupported Media Type) Error. It was tested with tcpdump and wireshark. After changing the return value to "{'content-type': self._parameters['content_type']}" of the get_request_headers method the problem was solved.