Support for ESRI Feature- en MapServices in 1-and-the-same probe called esri
The complete URL of the Feature- of MapService is the input for the probe. Based on the last part of the url (FeatureServer or MapServer) the code uses another json field to get the identityfield.
A drilldown test is niet possible in case of a query test because the url contains the number of the layer to test. Due to this the drilldown test gives an error.