Add run message and report to webhook message
This does increase the message size, though I imagine that's marginal,
and is backwards compatible in that old keys remain in place and
aren't modified. The goal is to make the actual GHC app more set and
forget, and we can handle the errors with a little more discretion on
the webhook end.

The report is being sent as URLEncoded form data, which if we send a
hierarchy just sends a single key. We want the entire report so encode
the JSON as a string and send that. The keys are sorted to make
testing easier.